Refactored introspection code
